cloud_functions/gmail_ingest/requirements.txt
requests>=2.31.0
google-cloud-firestore>=2.11.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.108.0
flask>=2.3.0
